+++ sdl.cpp
@@ -1,13 +1,13 @@
 internal void HandleKeyDown(SDL_Keysym key, InputState * GameInputState)
 {
-    u8 Index = key.sym&255;
+    u8 Index = (u8)(key.sym&255);
     GameInputState->KeyWasDown[Index]=GameInputState->KeyIsDown[Index];
     GameInputState->KeyIsDown[Index]=1;
 }
 
 internal void HandleKeyUp(SDL_Keysym key, InputState * GameInputState)
 {
-    u8 Index = key.sym & 255;
+    u8 Index = (u8)(key.sym & 255);
     GameInputState->KeyWasDown[Index]=GameInputState->KeyIsDown[Index];
     GameInputState->KeyIsDown[Index]=0;
 }
@@ -32,7 +32,7 @@ internal b32 SetupSDLWindow(GameState * CurrentGameState)
     SDL_GL_SetAttribute( SDL_GL_CONTEXT_MINOR_VERSION, 0 );
 
     SDL_GL_SetAttribute( SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_CORE );
-    SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE, 24); 
+    SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE, 24);
 
 
     CurrentGameState->MainSDLWindow = SDL_CreateWindow("TA_SDL: SOMETHING HERE!!!", SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED,CurrentGameState->ScreenWidth, CurrentGameState->ScreenHeight, SDL_WINDOW_SHOWN|SDL_WINDOW_OPENGL);
@@ -46,13 +46,13 @@ internal b32 SetupSDLWindow(GameState * CurrentGameState)
     Assert(CurrentGameState->PerformanceCounterFrequency);
 
     GLint GLMajorVer, GLMinorVer;
-    
+
     GLenum ErrorValue = glGetError();
     if(ErrorValue!=GL_NO_ERROR)
     {
 	LogError("failed before create context : %s", gluErrorString(ErrorValue));
     }
-    
+
 
     SDL_GLContext gContext = SDL_GL_CreateContext( CurrentGameState->MainSDLWindow);
     SDL_GL_MakeCurrent (CurrentGameState->MainSDLWindow,gContext);
@@ -71,9 +71,9 @@ internal b32 SetupSDLWindow(GameState * CurrentGameState)
     {
 	LogError("failed before glewInit : %s",gluErrorString(ErrorValue));
     }
-  
 
-    glewExperimental = GL_TRUE; 
+
+    glewExperimental = GL_TRUE;
     GLenum glewError = glewInit();
     if( glewError != GLEW_OK )
     {
@@ -81,12 +81,10 @@ internal b32 SetupSDLWindow(GameState * CurrentGameState)
 	return 0;
     }
 
-    
+
     //as we need to use glewExperimental - known issue (it segfaults otherwise!) - we encounter
     //another known issue, which is that while glewInit suceeds, it leaves opengl in an error state
     ErrorValue = glGetError();
 
     return 1;
 }
-
-
